Outdated caches were not recreated when file and wiki state match.
This is usually the case after exports, that file state matches
the wiki state (as it should be) but the cache still reflects the
file state before the last export. Those changes would then show
up as potential conflicts later on when file and wiki state would
no longer match, most often because the message was updated in
the wiki after the last export.

Reported by Siebrand.

Change-Id: I36adfa751eecf009f1ba32dc265861db023c5f19