Show the priority languages in Special:Translate.
Adding the priority language names to Special:Translate and showing
a proper message if the reason is empty.
The functions for showing the error and warning were very similar,
so I turned them into one call.
The reason, if it is provided, is put in <span> with matching
lang and dir attributes according to the content language.

Patch set 2:
* Comment spelling.
Patch set 3:
* Rebased.
* Improvements according to comments from Pau and Santhosh:
  * Reworded the messages for easier understanding.
  * Changed the params accordingly.
  * Emphasized the first line bold for easier reading.
  * Used errorbox and messagebox classes.
  * Not showing the reason at all if it's not provided.
Patch set 4:
* Changed "The translation administrator" to
  "A translation administrator".

Change-Id: I6c454aaca84c374d011841ec7f39237a7ccdeaf6