Show the priority languages in Special:Translate.
Adding the priority language names to Special:Translate and showing
a proper message if the reason is empty.
The functions for showing the error and warning were very similar,
so I turned them into one call.
The reason is put in <span> with matching lang and dir attributes
according to the content language.

Change-Id: I6c454aaca84c374d011841ec7f39237a7ccdeaf6