Update patch set 4
Patch Set 4:

As general comments, this class doesn't necessarily need to live inside Translate extension. In the beginning it might be easier to develop it when you can make needed changes to other classes too. This would be a good time to start splitting the huge Groups.php and MessageGroups.php to one file per class.

Patch-set: 4