Update patch set 3
Patch Set 3: Abandoned

Taking Nikerabbit's feedback on-board, I'll abandon this as a feature. That said, I'll be introducing a change to the TranslateSvg extension to achieve a similar effect via JavaScript, because I think it's really, really important that we don't confuse casual users with use of the term "message group" when they will have no idea what one of those is.

Patch-set: 3