Update patch set 8
Patch Set 8: No score

The later patch sets tried different ways to tackle the date format issue that Niklas pointed out. Finally, the solution I picked was to change the format in the datepicker itself and to add zeros. I tested it on the translatewiki sandbox and the chart seems to come out correctly.

Patch-set: 8