Update patch set 1
Patch Set 1:

There is a very long dependency chain, is this intentional?

Please use topic branches for each changeset (e.g. translatesvg-foo). If they need to depend on another change, that's fine. But if that's not the case, please rebase on master.

By the way, where is mw.hooks coming from? Is that from translate? It should be namespaced better since this will conflict with core.

Also, the way this "mw.hooks" thing is improved is unacceptable imho, has this been designed/specified somewhere?

Please document more. I'm not indicating -1 because I don't know the proper context, and there is too many unknown variables.

Patch-set: 1