Update patch set 2
Patch Set 2:

That alternative is a workaround for a specific instance of the problem this change aims to fix. The real issue is that the table is inflexible and won't play well with other elements of the page (and note that this is wikitext, not even unrestricted html).

Why does it *have* to span 100% of the page width? The layout of pages should be customizable by users, not dictated by the developers.

If it's absolutely required that the table spans 100% of the page width, it should at least have "clear:both;" rules in its CSS so it doesn't break the layout of the page as in the example I gave.

But unless there's a good reason for the inflexibility, I'd say having the table adapt to the available width is a good thing. Its content wraps cleanly anyway, and we could even introduce a "min-width" rule to ensure it doesn't get too small.

Patch-set: 2