Update patch set 1
Patch Set 1:

Could do, if wanted. Makes for less files to run to setup the tables. But the schema updates would just create it at the end of the database update run, so it would still get applied, as it'll check that the index doesn't exist, and then create it

Patch-set: 1