Update patch set 3
Patch Set 3: There's a problem with this change, please improve

Still has the selector escaping issue.

Quick googling lead me to this: http://stackoverflow.com/questions/739695/jquery-selector-value-escaping

It also assumes that every opened editor has matching table row (not the case), but fortunately if it doesn't find it, it will just ignore so that is not a problem.

Patch-set: 3