(bug 38306) Page translation reuses deleted section ids
Added code to keep track of highest used id to avoid reuse.
Also contains fix where you could not mark page again if
it was marked with 0 sections (another bug caused it to
check the last marked version, not was was saved). So
removed the check for number of sections altogether.
Now it is possible to create translatable pages with
no translatable sections besides page title.

Change-Id: I6947b9c3b735805d649f744844edb2e16ee76f7f