Provide more info to 'TranslateFormatMessageBeforeTable' hooks
Specifically, add $group and $targetLanguage parameters and
simplify by passing $m (which has translation() and original()
methods) rather than $translation and $original.

Found I needed these details when I came to tweak my own usage
of the hook, which is sufficiently new (and still in this branch)
that I can't believe anyone else is using it at the moment.

Change-Id: If063e4dddbbb0104dc58968527971f879283c30e