Update patch set 3
Patch Set 3: There's a problem with this change, please improve

Removing this rule will cause the table to become narrower if there are only a few translations. IMO this doesn't look good and should be prevented. 

And to be honest, I prefer the current layout, with the table on top and the text with floating logo below. This keeps the content (page text) neatly separated from the interface (the table).

Patch-set: 3