Update patch set 3
Patch Set 3:

Thanks, Matmarex, for noticing that issue. I agree that this change should be abandoned because of that, but I do think that cases such as the one present in http://meta.wikimedia.org/w/index.php?title=Wikidata&oldid=3911709 need to be prevented. I'll submit a follow-up patch set introducing a "clear:both" rule in the table's css that would avoid such a case where having 100% with would be problematic.

Patch-set: 3