Added framework for key generation algorithms in GettextFFS
This is the most common local mod I've done in various wikies.
The algo can be specified in YAML using key keyAlgorithm in
the FILES section.

Added simple as alternative algorithm, but kept the current one
as default for BC.

Also simplified parseGettextSection by moving some logic to
the caller. That avoids passing 5 (one more would have been added
in this commit) parameters to it.

Change-Id: I4bcd984e7e2d1fca63428c31326d45629b68738e