Update patch set 2
Patch Set 2:

The require( './SuperUser.php' ) can not work since the path in not in the autoloader, you should have prefixed it with __DIR__. Anyway, with patchset 2 I have added the class to the _autoload.php file.

I also made SuperUser to extend TestUser which has been introduced in core a few weeks ago.

Patch-set: 2