Update patch set 1
Patch Set 1:

And there is also $('.sortable').data('tablesorter').config.sortList, which informs you about how the table is currently sorted. (Unfortunately it's in the ugly "internal" format, but it's easy enough to grasp - first elem. of the arrays in the index, second is 0 for asc, 1 for desc).

Also, it is possible to sort a table based on multiple columns, by holding the Shift key while clicking – the .sortList will reflect this. This should probably be handled here.

Patch-set: 1