Update patch set 2
Patch Set 2: There's a problem with this change, please improve

This fails because Jenkins cannot install deps at the moment, and your test expects language names in English.

You should probably check if CLDR is installed, and if it's not, skip the test, or create two tests, and execute the one that's "right" depending on CLDR being available or not.

Patch-set: 2