Reduce queries in WorkflowStatesMessageGroup
With TranslateUtils::getContents we can load all definitions
with one query. We can then loop over the results to see
if any are missing, avoiding the queries caused by Title::
exists().

Also changes it to gather states from all groups, not just
from the default configuration.

The creation of missing state pages should use jobs to avoid
recursion issues, but that is not done in this commit.

Change-Id: If8bffc93903d5a77d93c90187c4da83c5cd347d0