Update patch set 5
Patch Set 5: There's a problem with this change, please improve

I fixed all the issues that Niklas pointed out, except "Is the list emptied for every keyup event?".

I tried addressing this in patch set 2, with lastSearchQuery, but Santhosh removed it in patch set 3, and I'm not sure why.

Patch-set: 5