WIP: adding display and dynamic loading of warnings
MessageChecker function copied from ext.translate.editor.js
with modifications.

In the old editor the TranslationHelpers sends warnings as
formatted HTML. For TUX I made it send them as a JSON list.
It may be a bit hacky and improvement suggestions are welcome.

Some CSS improvements are still needed, too.

Change-Id: Iec2c6c6c00e453c2caae08940c114d4c48b66339