Add appropriate lang and dir to documentation
In practice, this change is not supposed to affect the display
to users of left-to-right languages. To users of right-to-left
this change improves the display, because without it
the docs in English appear right-to-left.
The row classes override the alignment settings
and I'm not sure how to fix it nicely, so the display
is not perfect - the margins are too narrow. But it's
still better than seeing English right-to-left.

Change-Id: I488c6c5b25f23be9ec1f4d6a51f8ad678dd65879