Fix hook signature (follow-up to I8b6a75a3).
Use separate entry points for the old EditFilterMerged and the new EditFilterMergedContent
hooks, since their signature differ quite a bit.

This also fixes the actual error reporting, which was broken for
the new hook.

Note: This is UNTESTED, since i don't know whow to set up Translate for testing this.

Change-Id: I896d08b49706f96af525770816f08ff68a70d04e