Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

It looks nice. 

The bigger problem with the inline editor is deciding whether to provide a way to access the full editor for messaged that have a long description a with lot of wikitext. Depending on how realistic is that escenario we can find a solution with the appropriate emphasis level: allow to open the link in new window with the full editor, adding a small icon to open in new window, add a text link next to the cancel button..

Other minnor UX comments to consider:
* Text area should have overflow:auto to allow long descriptions to be scrollable
* The "Edit the documentation" label can be ommited and use the textarea placehlder instead (since it is mainly useful for new messages more than existing ones)
* Save button should be enabled only when descriptin is modified.
* After clicking on edit, the input focus should move to the text area.

Patch-set: 3