Make action bar to float when needed
The action bar may be hidden by long list of messages (especially when those are loaded dynamically).
this change makes the action bar to float when it is hidden by scroll so it becomes visible.
When the user reaches the end of the list (or for short list of messages that fit the current viewport), the action bar is positioned at the bottom of the message list.

Change-Id: I148f6ba130efa80cfd3a2bf28c8d4e9467faa435