Adjust z-index for action bar to avoif overlapping with message editor
The z-index has been added since the text area of the message editor
was displaying on top of the action bar while the rest of the editor
was displayed below (in terms of depth). As a result, depth was inconsistent.
with this change, the action bar is displayed on top of the editor when floating.

Change-Id: I91a6e8ed62fa3f7fb55a2a389822d77480a84d65