Update patch set 1
Patch Set 1: There's a problem with this change, please improve

I've been checking the patchset. I think I prefer not adding it as a separate link until we see if more editing support is needed. Since the edit box becomes larger we are probably offering enough space for comfortably editing. Later, if we see that it is needed fo one of the text areas, or both, we can think on a more global solution.

Making the "Edit translation" link able to be opened in a new window  would provide access to the functionality for advanced users without adding complexity to the UI. But I would prefer to see fixed some of the issues we already detected users are facing before adding other extras.

I also found that the patchset breaks the documentation shortening and expansion for large documentation text (probably due some class renaming).

Patch-set: 1