Update patch set 2
Patch Set 2: There's a problem with this change, please improve

Rebased, but either I broke something or it doesn't quite work as expected, it is not appearing always.

Also this is quite spaghetti, should likely be moved to messagetable and use hooks/events or something to make all be in one place.

Patch-set: 2