Update patch set 1
Patch Set 1:

Messages not cleared during group or filter changes is one reason for this bug. I could reproduce and stop using this patchset. 

random response arrival on mode switching will not be a problem since it is same messages we are presenting in different ways.

Bug random order response from group switching or filter switching is an edge case to handle properly. but not planning to do a fix for that in this commit. We will need to abort existing xhrs if any while switching.

Patch-set: 1