Update patch set 1
Patch Set 1:

I tried this but still prefer the idea of a "Loading..." label. The reason is that the loading indicator attracts visual attention to communicate that you should wait. This makes sense when waiting for infinite scrolling to load new messages.
In the case of the helper area I would prefer a very subtle indicator that does not distract you from the translation area, but in case you are looking for more info you are not skipping the translation until information is presented.

Patch-set: 1