Update patch set 1
Patch Set 1:

I added the icon but this requires a change in the current HTML structure:
The "translated-by-self" element should be moved to be the first child inside "tux-proofread-action-block".

I would have done this myself, but all these nodes are crated by nested appends which makes it hard to insert something in the middle without major changes in the patchset.

Patch-set: 1