Update patch set 5
Patch Set 5:

Tested with Pau. For now we decided to show the button immediately when the user starts typing and not after deleting, as the bug says. We may change it later.

In the latest patch set I fixed the problem with messages remaining "Unsaved" after discarding the changes.

Patch-set: 5