Update patch set 5
Patch Set 5: Code-Review-1

I've tested this, and I don't think it's up to spec yet.

Expected result: The selector open immediately and gives feedback that data is being loaded.

I don't see this happening. I still have to wait for <something>. I start clicking "All" as soon as it's visible, and it's waiting for a backend call to finish before actually displaying the group selector.

Patch-set: 5