Update patch set 1
Patch Set 1:

eg: For a translate message list item - we are binding an editor. the editor is appended to the message div. If the div has click event to open editor, closing the editor will trigger bubble the event to container, opening the editor again. There are multiple ways to prevent this happening. return false is simple approach that jquery respect from handlers.

Patch-set: 1