Update patch set 5
Patch Set 5: Code-Review-1

Works as advertised, but differs a bit from the specification. The spec has the text "x% translated, y% proofread, z% to review". "z% to review" does not make sense to me. I've asked Pau to clarify.

Also, I see that there is no fuzzy percentage. I'd propose to omit fuzzy if there is no fuzzy, and add a second message "$1% translated, $2% proofread, $3% outdated" which is only used when there is a certain percentage outdated. When there are no outdated messages in the group, the current message should remain to be used. Should we add that now or later?

Patch-set: 5