Speed up translatable page move page
Applied two optimizations here:
* Avoid adding lines one by one - add them in batches. Each line
  is wikitext and starting parser thousands of times is both slow
  and takes 100M+ of memory.
* Do not validate all of the section pages for move, only one.
  Validating one page causes 5 sql queries on my wiki and I didn't
  see any obvious way to preload that information. A compromise
  between unlikely errors and speed.

Tested with translatable page which has 2500 related pages, it went
from 30s (timeout) and lots of memory to (with profiling enabled)
to 2s and 50M of memory (profiling still enabled).

It is not clear whether this completely fixes bug 41138. The original
report refers to timeout before actually executing the move, but some
comments hint that this would happen after the verification step,
which would be totally different issue.

Bug: 41138
Change-Id: I8ec9bf3e9ed977e48584aa37bc9949e2fbcabb24