Update patch set 2
Patch Set 2:

If there are multiple suggestions, it seems like the last one will always be inserted.

Another issue that I found: When selecting text, to copy it for example, this is not possible. I select the text, and as soon as I let the mouse button go, the text is unselected and the insert is performed.

Patch-set: 2