Update patch set 3
Patch Set 3: Code-Review-1

This doesn't result in expected behavior of steps 4 and 6 in IE10. The observed behavior is still as documented. Please ensure it is tested with the URL in the bug report, as Niklas reported to have seen improvement in another scenario.

Patch-set: 3
Reviewer: Siebrand <92@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1