Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Also, in general: I'd recommend commenting the line out in PHP instead of removing it. There is no harm in the attribute being there if it is not supported. In fact, it should stay there so that if there is a polyfill script on the page, it can pick it up and make it work in old browsers.

The reason you do this is because the polyfill in core has a bug you want to avoid temporarily. That is understandable, but be sure to document that because now it looks like you're removing it for old browsers for other reasons.

Patch-set: 1
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1