Update patch set 2
Patch Set 2:

There is a reason to make it so short. Afaik it's used on Special:Translations as a super scripted prefix. Making that longer, will make it look worse. I'm not aware of additional use cases.

I'm not for merging this change. Would be interested in Pau's opinion, although this pages should not be micro optimized, but redesigned.

Patch-set: 2