Update patch set 1
Patch Set 1:

Krinkle: in practice this seems to only happen when mouse cursor is already over an element when the $( document ).ready() call binds on the events. That is why simple global check is enough to cater for the missing hover-in (which is either not send at all or send before we bind to the event - I do not which and it doesn't matter). Perhaps rewording the commit summary to be more exact would help?

Patch-set: 1
Label: Verified=0