Update patch set 2
Patch Set 2:

I'm not sure why need commitwithin if we've got autocommit.  One or the other should be fine.  Removing the explicit commits is fine so long as we have a way to force a rebuild of the whole index or just the last few seconds in the case of a solr crash.  Or maybe we just don't care in this case.

Patch-set: 2
Label: Verified=0