Update patch set 1
Patch Set 1:

IIRC, the full screen mode is supposed to come just after the message filters starts floating on the screen. A click on expand icon near wiki logo is not mandatory.

And the expand icon is not reachable just after user start scrolling the page. So that functionality is hidden from user.

Pau can confirm it.

Patch-set: 1