Update patch set 1
Patch Set 1:

Is easier for debugging in test setups (like looking at the objectcache table) and when looking at error logs if the keys are not just one hash.

If in bagostuff, things like "wfDebugLog( 'memcached-serious', $msg );" in the pecl client class would need to use the pretty key but the pecl binding calls would need to use the hash key. This could be done, but it would mean the shorting function would be sprinkled in the subclasses. It would have been easier if there were "final" delegating functions...

It would be nice if BagOStuff could not bother sending overly-long keys over the wire and error out.

Patch-set: 1
Reviewer: Aaron Schulz <16@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0