Allow text selection in page mode
Bind the click event handler to one level deeper: not the element
which has both the message display and editor, but to the message
display only.

Replaced return false with e.preventDefault to allow clicks to
bubble up and close dialogs for example. This only blocks for
example opening of links rendered in messages. And even those
can still be accessed via the context menu.

Bug: 47861
Change-Id: I941b7a9ed0ba4672b9dc71bb706c225ba5d9c712