Make jquery.autosize aware of style changes
The jquery.autosize plug-in is used to dynamically expand or shrink the
textarea while translating. Once it is initialized for a textarea, and
if no other textarea is edited in-between, it will not copy all styles
again to its hidden mirror-textarea that is used to compute the required
size of the textarea for performance reasons.

This works fine as long as the explicit and inherited styles of a
textarea do not change. But if they do, the plug-in gets it wrong and
consequently applies a wrong size. This is a serious issue if the
textarea is smaller than its content because the plug-in also sets
overflow to hidden.

In the screenshot provided as attachment #3 in bug 48817, the font size
is 16px in the visible textarea and 13px in the mirror due to this
issue.

Bug: 48817
Change-Id: I9c4b2c928b181e4f3bac7a57109032c1d48b511f