Update jquery.autosize
So one can make it aware of style changes.

The jquery.autosize plug-in is used to dynamically expand or shrink the
textarea while translating. Once it is initialized for a textarea, and
if no other textarea is edited in-between, it will not copy all styles
again to its hidden mirror-textarea that is used to compute the required
size of the textarea for performance reasons.

This works fine as long as the explicit and inherited styles of a
textarea do not change. But if they do, the plug-in gets it wrong and
consequently applies a wrong size. This is a serious issue if the
textarea is smaller than its content because the plug-in also sets
overflow to hidden.

In the screenshot provided as attachment #3 in bug 48817, the font size
is 16px in the visible textarea and 13px in the mirror due to this
issue.

This is a slightly modified approach of
29880d5b55c3f6b0c4cc564b7b92bc131a568481

Doing
 $( 'textarea' ).trigger( 'autosize' ); or better
 $( 'textarea' ).trigger( 'autosize.resizeIncludeStyle' );
is now sufficient and does both: Resetting the mirror-textarea and
resizing the set of textareas it was called on.

Bug: 48817
Change-Id: I9049e8479c8a77d42cdd6d8c6909a30da8005f3c