Make Special:AggregateGroups using jquery.chosen
With the time, the number of translatable pages will grow. There are
already about 350 pages at Meta marked for translation.
This leads to a messy UI with the effect that you spend a lot of time
searching but not finding.
Using jQuery.chosen will ease finding the page you'd like to add to an
aggregated group.

jQuery.async is used to avoid freezing the browser and throwing
warnings about unresponsive scripts at the user.

Only a remote-data-service would solve this issue properly.

Bug: 52216
Change-Id: I0dc4b71814eaad82246bfb69d793bc2a48551895