Add @codingStandardsIgnoreFile in alias files
PH CodeSniffer raises a warning when lines are over 100 characters and an
error when a line is over 120 characters. Files with special page aliases
are generated using the magic-export.php script of the Translate extension
and can be "edited" using Special:AdvancedTranslate. There is no need to
have these files adhere to the convention. To not require any changes to
the PHP CodeSniffer configuration, Antoine Musso advised to use this
method to not raise unneeded warnings and errors.

Change-Id: Ic5f0dcfbc0f9fea1189a0bf6db348e85651948cf