Update patch set 7
Patch Set 7:

This version should be working, but apart from the unit tests I have not tested it.

The usage of the xmlns seems inconsistent around the example files I can find. Currently this code only works if it is not present. Do you know if what will cause problems? If not, I can avoid the complexity of dealing with files with and without it.

Patch-set: 7