Update patch set 7
Patch Set 7:

I spent some time looking for an xsi:schemaLocation xsd file that could be used for validation purposes and was not able to find one.

The problem with removing it, in the XML sax transformation, is that the transformation will look for message elements in this namespace.

So there may definitely break some things when this namespace declaration would be removed from the file.

Patch-set: 7