Update patch set 2
Patch Set 2:

The problem isn't this patch but other issues mentioned on the bug. I don't have time to work on core parser issues on my free time. We can merge this but before the aforementioned issues are fixed this option is not useful except perhaps for some developers working on those issues.

Patch-set: 2