Tests for accepting and rejecting users
Fixed a curious bug when manifested when we tried to recreate an
user with the same name who we had just deleted during the same
request. Since it is not clear whom to blame for this weird
behavior, I just worked around the problem by adding a lots of of
comments to the code.

Change-Id: I735010eb7f3af81d4072a2f44fa5cdf9caf26a3c