Promotion emails and reminders
Refactor the reminder code to be able to send both reminders
and promotion emails.

Changes the 'remind' API so that 'body' and 'subject'
params are removed, and only a constant canned reminder is sent.

Removes the reminder dialog - it was just an experiment that was never used, and it's hard to use it to send localized emails.

Implement smart displaying of reminder count with human timestamp.

Change-Id: Ie81c50787b5b5ed2e711631ecceef4008ec44344