Prevent disappearance of user details after using ULS
There was an issue with using ULS: when selecting a language
into which nobody translates and then clearing the filter,
the usernames would not appear.

This is fixed by adding the 'pane' class when emptying
elements with the 'details' class and renaming the other
element with 'details' class to 'request-info'.

Tests included.

Change-Id: Iff87c7db07ab4f3a86b8fff4fd7d2b2f8e28f980