Update patch set 2
Patch Set 2: Code-Review-1

The root cause for this issue is the class 'details' used for two purpose - for username+useremail fields and for showing the request details.

So when you says $('.details').empty() it clears both even though you wanted to empty one of them. The fix here was using another class 'pane' for request details and using it as '.details.pane'. More than this, the 'pane' class is also used with requests. 

This is too confusing and in future one will use 'details' class for referring one and will go through this issue. 

Solution here is, use one class name for one and only one purpose. Violating that was the root cause for this issue and that will potentially create issues in future.

In this specific case, you may rename the 'details' class for request details as 'request-details' or 'request-details-pane' and get rid of all .pane classes because there  is no css defined for that class or that class alone cannot be used to refer any elements in the page. There is a set of .panes in the css - search.pane, filter.pane, requests.pane, details.pane

Patch-set: 2
Reviewer: Santhosh <110@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1