Update patch set 4
Patch Set 4:

I can reproduce "Element is no longer attached to the DOM" error message on my machine in about 10% of test runs, the remaining 90% of the time the test runs just fine. I am not sure what the problem is. It might be a bug in Selenium Ruby bindings or page-object gem. I will try to reproduce the bug with as little code as possible and report it to Selenium and page-object bug tracker.

Patch-set: 4
Reviewer: Zfilipin <487@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0