Update patch set 1
Patch Set 1:

Translate and the extensions part of MLEB are supported up to two stable versions of MediaWiki.

The same also goes for extensions part of SemanticBundle, although they may even be going back to the latest LTS version. We could consider introducing a TranslateSpecialPage in which this method is overridden, similar to what you did in AbuseFilter.

Patch-set: 1