Update patch set 1
Patch Set 1:

I hardcoded the additional colon, but I wonder if there is a dedicated function in Title.php for that (I didn’t find).

There is an other solution of changing all existing translations to add the colon, but this is more error-prone and less intuitive for translators.

The message MediaWiki:Translate-tag-page-desc currently adds a prefixed colon of the second argument of [[$2|$1]]. This could be shortened by having only one (correctly prefixed) argument, but this make touching all existing translations. So I didn’t include it in this patch, but it could.

Patch-set: 1
Label: Verified=0