Statsbars had incorrect values for subgroups
The cause was that we were not passing all options to the subgroup
selector instances. Language was missing, so it took the statistics
for English language as default value. Fixed by passing all explicit
options to the instances. For this I had to store those separately
from the general options which were merged with default values, or
the repositioning of the menu to match the new sub-item in the group
trail would not work.

Bug: 60128
Change-Id: I271cf262c360d074cdad339cfa334a99562abf8a