Display outdated page translations as such, without removing them
It has always been the case in Translate and users ask for it loudly.
This partially reverts commit daa576affc38dd0e9179d97b774a36df92f60edf.

This patch adds a preference $wgTranslateKeepOutdatedTranslations (true
by default). If this preference is set to false, then this patch should
be a no-op compared to current version.

Bug: T60429
Change-Id: I43aa0ec577d2fcd40418348ee430f7594935651d