Improve Special:LanguageStats purging
As mentioned in bug 42634 comment #3, this greatly improves the
expected results of action=purge on Special:LanguageStats and
Special:MessageGroupStats. Previously action=purge would wipe out
the message group stats cache, but not the display rows which are
cached separately. Fixed this by moving message group stats cache
purging to happen after re-rendering the display rows. The cache
key of display rows contain the stats, so we were only purging
the display rows for the "stats not yet calculated" case.

Change-Id: I0d95f974d0e9c5ff45430f9f1f7b16cb5227b09c