Simplistic alignment based on h2 headers for Special:PageMigration
h2 headers are aligned in the order they appear in the source and target
units. If needed, blank units are added and if there is no room for aligning,
the preceding units are merged until there is enough room for alingment.
QUnit tests have also been added to verify the same.

Bug: 66162
Change-Id: I06fea8b76b01014a83c32b972b15375aa8f44e1a