Update patch set 11
Patch Set 11:

(6 comments)

With few lines of JavaScript, you could hide/show the title section when the checkbox is checked or unchecked for nicer UI. That will be easier if you add a class to the html, see a inline comment.

If you prefer, I can also merge this first, once the other comments have been addressed, and you can do the JavaScript part in separate commit.

Patch-set: 11
Label: Verified=0