Update patch set 13
Patch Set 13:

If the page title is initially marked for translation and later removed and there is a translation available for the page title, it shows up the translation.

Does it need to be addressed? Because this patch aims to save translators' time by not showing them the title for translation, so if it shows the translation (if available) on existing wikis, then it should be ok.

Nikerabbit, Siebrand- comments?

Patch-set: 13