Use json rather than simplejson
I don't have hard numbers to back this up, but since json is included by
default with Python 2.6+, I wager it is more common than the standalone
simplejson module, to which it is functionally equivalent. So use it instead.

Also fix a small typo (extented -> extended) which occurred twice in the code.

Change-Id: Ie88de54579f0081b8933156309f1ed868d4c6d31