Special page Special:PagePreparation for translate page preparation
The script ext.translate.pagepreparation.js is now invoked on a button click
at Special:PagePreparation by taking the page name as input from user. The diff
between the current revision and the prepared text is shown. The user can then
click on the 'Save' button to Save the page with the prepared text. A link to edit
the page is provided after saving is done.

Change-Id: I42a06acc6cc15c3c06bb74964a525eacca910a3b