Update patch set 3
Patch Set 3:

I had written a comment which I forgot to publish. Here it is:

I don't think this code belongs here. This special page already handles updating of translation pages.

The issues is that if someone moves translation unit pages directly, and when they do that they are not using this special page.

Patch-set: 3
Label: Verified=0