Update patch set 3
Patch Set 3: Code-Review-1

Thanks for the explanation. This is ok for me, but Niklas will veto any code executed by default which assumes the existence of a template.

I suggest that you file a bug, blocker of this one, for the creation of {{translation}} as a magic word (initially in Translate). Time seems mature now.

Patch-set: 3
Label: Code-Review=-1