'#translation' added as magic word in Translate extension
One cannot rely on the presence of {{translation}} on the wiki, hence
a magic word has been created to have the same behavior as that of the
template. This has been accomplished by using the magic word as a parser
function. The function checks for the language code in the page title,
and returns it if it is a valid code. Else, it returns an empty string,
meaning that it is not a translation page.

Bug: 68385
Change-Id: I16bb4fe5294c238aa345e9b33350d64d518748f4